Adjusts interactions with ManifestFinderRegistry
* Extracts `IManifestFinderRegistry` interface
* Moves reflection-based registration to another class,
`ManifestFinderRegistryLoader`
