+++ Corgibytes.Freshli.Lib/IManifestFinderRegistry.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Corgibytes.Freshli.Lib
+{
+    public interface IManifestFinderRegistry
+    {
+        IList<AbstractManifestFinder> Finders { get; }
+
+        void Register(AbstractManifestFinder finder);
+    }
+}
