+++ Corgibytes.Freshli.Lib/ManifestFinderRegistry.cs
@@ -6,7 +6,7 @@ using NLog;
 
 namespace Corgibytes.Freshli.Lib
 {
-    public class ManifestFinderRegistry
+    public class ManifestFinderRegistry : IManifestFinderRegistry
     {
         private readonly Logger logger = LogManager.GetCurrentClassLogger();
 
@@ -19,45 +19,6 @@ namespace Corgibytes.Freshli.Lib
         {
             Finders.Add(finder);
         }
-
-        public void RegisterAll()
-        {
-            var manifestFinderTypes = new HashSet<Type>();
-            foreach (var assembly in AppDomain.CurrentDomain.GetAssemblies())
-            {
-                foreach (var type in FindersLoadedIn(assembly))
-                {
-                    manifestFinderTypes.Add(type);
-                }
-            }
-
-            foreach (var type in manifestFinderTypes)
-            {
-                logger.Log(
-                    LogLevel.Info,
-                    $"Registering AbstractManifestFinder: {type}"
-                );
-                Register((AbstractManifestFinder)Activator.CreateInstance(type));
-            }
-        }
-
-        private IEnumerable<Type> FindersLoadedIn(Assembly assembly)
-        {
-            try
-            {
-                return assembly.GetTypes().
-                    Where(
-                        type => type.BaseType == typeof(AbstractManifestFinder) &&
-                                type.GetConstructor(Type.EmptyTypes) != null
-                    );
-            }
-            catch
-            {
-                logger.Log(LogLevel.Info, $"Unable to load types from {assembly}");
-                return new List<Type>();
-            }
-        }
-
-
     }
+
 }
